Update booking to print result

+++ src/nuber/students/BookingResult.java
@@ -2,7 +2,7 @@ package nuber.students;
 
 public class BookingResult {
 
-	public long jobID;
+	public int jobID;
 	public Passenger passenger;
 	public Driver driver;
 	public long tripDuration;
@@ -13,7 +13,7 @@ public class BookingResult {
 	 * @param driver
 	 * @param tripDuration
 	 */
-	public BookingResult(long jobID, Passenger passenger, Driver driver, long tripDuration)
+	public BookingResult(int jobID, Passenger passenger, Driver driver, long tripDuration)
 	{
 		this.jobID = jobID;
 		this.passenger = passenger;
